+++ core/src/main/scala/japgolly/scalajs/react/vdom/ReactTags.scala
@@ -1,5 +1,6 @@
 package japgolly.scalajs.react.vdom
 
+import japgolly.scalajs.react.ReactElement
 import scalatags._
 import generic.Util
 
@@ -7,7 +8,7 @@ import generic.Util
  * Trait that contains the contents of the `Tags` object, so they can be mixed
  * in to other objects if needed.
  */
-trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
+trait ReactTags extends generic.Tags[VDomBuilder, ReactElement, ReactFragT]{
 
   // Root Element
   /**
@@ -16,7 +17,7 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val html = "html".tag[ReactOutput]
+  val html = "html".tag[ReactElement]
 
   // Document Metadata
   /**
@@ -25,26 +26,26 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val head = "head".tag[ReactOutput]
+  val head = "head".tag[ReactElement]
 
   /**
    * Defines the base URL for relative URLs in the page.
    *
    *  MDN
    */
-  val base = "base".voidTag[ReactOutput]
+  val base = "base".voidTag[ReactElement]
   /**
    * Used to link JavaScript and external CSS with the current HTML document.
    *
    *  MDN
    */
-  val link = "link".voidTag[ReactOutput]
+  val link = "link".voidTag[ReactElement]
   /**
    * Defines metadata that can't be defined using another HTML element.
    *
    *  MDN
    */
-  val meta = "meta".voidTag[ReactOutput]
+  val meta = "meta".voidTag[ReactElement]
 
 
   // Scripting
@@ -54,7 +55,7 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val script = "script".tag[ReactOutput]
+  val script = "script".tag[ReactElement]
 
 
   // Sections
@@ -64,58 +65,58 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val body = "body".tag[ReactOutput]
+  val body = "body".tag[ReactElement]
 
   /**
    * Heading level 1
    *
    *  MDN
    */
-  val h1 = "h1".tag[ReactOutput]
+  val h1 = "h1".tag[ReactElement]
   /**
    * Heading level 2
    *
    *  MDN
    */
-  val h2 = "h2".tag[ReactOutput]
+  val h2 = "h2".tag[ReactElement]
   /**
    * Heading level 3
    *
    *  MDN
    */
-  val h3 = "h3".tag[ReactOutput]
+  val h3 = "h3".tag[ReactElement]
   /**
    * Heading level 4
    *
    *  MDN
    */
-  val h4 = "h4".tag[ReactOutput]
+  val h4 = "h4".tag[ReactElement]
   /**
    * Heading level 5
    *
    *  MDN
    */
-  val h5 = "h5".tag[ReactOutput]
+  val h5 = "h5".tag[ReactElement]
   /**
    * Heading level 6
    *
    *  MDN
    */
-  val h6 = "h6".tag[ReactOutput]
+  val h6 = "h6".tag[ReactElement]
   /**
    * Defines the header of a page or section. It often contains a logo, the
    * title of the Web site, and a navigational table of content.
    *
    *  MDN
    */
-  val header = "header".tag[ReactOutput]
+  val header = "header".tag[ReactElement]
   /**
    * Defines the footer for a page or section. It often contains a copyright
    * notice, some links to legal information, or addresses to give feedback.
    *
    *  MDN
    */
-  val footer = "footer".tag[ReactOutput]
+  val footer = "footer".tag[ReactElement]
 
 
   // Grouping content
@@ -124,81 +125,81 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val p = "p".tag[ReactOutput]
+  val p = "p".tag[ReactElement]
   /**
    * Represents a thematic break between paragraphs of a section or article or
    * any longer content.
    *
    *  MDN
    */
-  val hr = "hr".voidTag[ReactOutput]
+  val hr = "hr".voidTag[ReactElement]
   /**
    * Indicates that its content is preformatted and that this format must be
    * preserved.
    *
    *  MDN
    */
-  val pre = "pre".tag[ReactOutput]
+  val pre = "pre".tag[ReactElement]
   /**
    * Represents a content that is quoted from another source.
    *
    *  MDN
    */
-  val blockquote = "blockquote".tag[ReactOutput]
+  val blockquote = "blockquote".tag[ReactElement]
   /**
    * Defines an ordered list of items.
    *
    *  MDN
    */
-  val ol = "ol".tag[ReactOutput]
+  val ol = "ol".tag[ReactElement]
   /**
    * Defines an unordered list of items.
    *
    *  MDN
    */
-  val ul = "ul".tag[ReactOutput]
+  val ul = "ul".tag[ReactElement]
   /**
    * Defines an item of an list.
    *
    *  MDN
    */
-  val li = "li".tag[ReactOutput]
+  val li = "li".tag[ReactElement]
   /**
    * Defines a definition list; al ist of terms and their associated definitions.
    *
    *  MDN
    */
-  val dl = "dl".tag[ReactOutput]
+  val dl = "dl".tag[ReactElement]
   /**
    * Represents a term defined by the next dd
    *
    *  MDN
    */
-  val dt = "dt".tag[ReactOutput]
+  val dt = "dt".tag[ReactElement]
   /**
    * Represents the definition of the terms immediately listed before it.
    *
    *  MDN
    */
-  val dd = "dd".tag[ReactOutput]
+  val dd = "dd".tag[ReactElement]
   /**
    * Represents a figure illustrated as part of the document.
    *
    *  MDN
    */
-  val figure = "figure".tag[ReactOutput]
+  val figure = "figure".tag[ReactElement]
   /**
    * Represents the legend of a figure.
    *
    *  MDN
    */
-  val figcaption = "figcaption".tag[ReactOutput]
+  val figcaption = "figcaption".tag[ReactElement]
   /**
    * Represents a generic container with no special meaning.
    *
    *  MDN
    */
-  val div = "div".tag[ReactOutput]
+  val div = "div".tag[ReactElement]
 
   // Text-level semantics
   /**
@@ -206,76 +207,76 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val a = "a".tag[ReactOutput]
+  val a = "a".tag[ReactElement]
   /**
    * Represents emphasized text.
    *
    *  MDN
    */
-  val em = "em".tag[ReactOutput]
+  val em = "em".tag[ReactElement]
   /**
    * Represents especially important text.
    *
    *  MDN
    */
-  val strong = "strong".tag[ReactOutput]
+  val strong = "strong".tag[ReactElement]
   /**
    * Represents a side comment; text like a disclaimer or copyright, which is not
    * essential to the comprehension of the document.
    *
    *  MDN
    */
-  val small = "small".tag[ReactOutput]
+  val small = "small".tag[ReactElement]
   /**
    * Strikethrough element, used for that is no longer accurate or relevant.
    *
    *  MDN
    */
-  val s = "s".tag[ReactOutput]
+  val s = "s".tag[ReactElement]
   /**
    * Represents the title of a work being cited.
    *
    *  MDN
    */
-  val cite = "cite".tag[ReactOutput]
+  val cite = "cite".tag[ReactElement]
 
   /**
    * Represents computer code.
    *
    *  MDN
    */
-  val code = "code".tag[ReactOutput]
+  val code = "code".tag[ReactElement]
 
   /**
-   * Subscript tag[ReactOutput]
+   * Subscript tag[ReactElement]
    *
    *  MDN
    */
-  val sub = "sub".tag[ReactOutput]
+  val sub = "sub".tag[ReactElement]
   /**
    * Superscript tag.
    *
    *  MDN
    */
-  val sup = "sup".tag[ReactOutput]
+  val sup = "sup".tag[ReactElement]
   /**
    * Italicized text.
    *
    *  MDN
    */
-  val i = "i".tag[ReactOutput]
+  val i = "i".tag[ReactElement]
   /**
    * Bold text.
    *
    *  MDN
    */
-  val b = "b".tag[ReactOutput]
+  val b = "b".tag[ReactElement]
   /**
    * Underlined text.
    *
    *  MDN
    */
-  val u = "u".tag[ReactOutput]
+  val u = "u".tag[ReactElement]
 
   /**
    * Represents text with no specific meaning. This has to be used when no other
@@ -284,20 +285,20 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val span = "span".tag[ReactOutput]
+  val span = "span".tag[ReactElement]
   /**
    * Represents a line break.
    *
    *  MDN
    */
-  val br = "br".voidTag[ReactOutput]
+  val br = "br".voidTag[ReactElement]
   /**
    * Represents a line break opportunity, that is a suggested point for wrapping
    * text in order to improve readability of text split on several lines.
    *
    *  MDN
    */
-  val wbr = "wbr".voidTag[ReactOutput]
+  val wbr = "wbr".voidTag[ReactElement]
 
   // Edits
   /**
@@ -305,13 +306,13 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val ins = "ins".tag[ReactOutput]
+  val ins = "ins".tag[ReactElement]
   /**
    * Defines a removal from the document.
    *
    *  MDN
    */
-  val del = "del".tag[ReactOutput]
+  val del = "del".tag[ReactElement]
 
   // Embedded content
   /**
@@ -319,79 +320,79 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val img = "img".voidTag[ReactOutput]
+  val img = "img".voidTag[ReactElement]
   /**
    * Represents a nested browsing context, that is an embedded HTML document.
    *
    *  MDN
    */
-  val iframe = "iframe".tag[ReactOutput]
+  val iframe = "iframe".tag[ReactElement]
   /**
    * Represents a integration point for an external, often non-HTML, application
    * or interactive content.
    *
    *  MDN
    */
-  val embed = "embed".voidTag[ReactOutput]
+  val embed = "embed".voidTag[ReactElement]
   /**
    * Represents an external resource, which is treated as an image, an HTML
    * sub-document, or an external resource to be processed by a plug-in.
    *
    *  MDN
    */
-  val `object` = "object".tag[ReactOutput]
+  val `object` = "object".tag[ReactElement]
   /**
    * Defines parameters for use by plug-ins invoked by object elements.
    *
    *  MDN
    */
-  val param = "param".voidTag[ReactOutput]
+  val param = "param".voidTag[ReactElement]
   /**
    * Represents a video, and its associated audio files and captions, with the
    * necessary interface to play it.
    *
    *  MDN
    */
-  val video = "video".tag[ReactOutput]
+  val video = "video".tag[ReactElement]
   /**
    * Represents a sound or an audio stream.
    *
    *  MDN
    */
-  val audio = "audio".tag[ReactOutput]
+  val audio = "audio".tag[ReactElement]
   /**
    * Allows the authors to specify alternate media resources for media elements
    * like video or audio
    *
    *  MDN
    */
-  val source = "source".voidTag[ReactOutput]
+  val source = "source".voidTag[ReactElement]
   /**
    * Allows authors to specify timed text track for media elements like video or
    * audio
    *
    *  MDN
    */
-  val track = "track".voidTag[ReactOutput]
+  val track = "track".voidTag[ReactElement]
   /**
    * Represents a bitmap area that scripts can use to render graphics like graphs,
    * games or any visual images on the fly.
    *
    *  MDN
    */
-  val canvas = "canvas".tag[ReactOutput]
+  val canvas = "canvas".tag[ReactElement]
   /**
    * In conjunction with area, defines an image map.
    *
    *  MDN
    */
-  val map = "map".tag[ReactOutput]
+  val map = "map".tag[ReactElement]
   /**
    * In conjunction with map, defines an image map
    *
    *  MDN
    */
-  val area = "area".voidTag[ReactOutput]
+  val area = "area".voidTag[ReactElement]
 
 
   // Tabular data
@@ -400,61 +401,61 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val table = "table".tag[ReactOutput]
+  val table = "table".tag[ReactElement]
   /**
    * The title of a table.
    *
    *  MDN
    */
-  val caption = "caption".tag[ReactOutput]
+  val caption = "caption".tag[ReactElement]
   /**
    * A set of columns.
    *
    *  MDN
    */
-  val colgroup = "colgroup".tag[ReactOutput]
+  val colgroup = "colgroup".tag[ReactElement]
   /**
    * A single column.
    *
    *  MDN
    */
-  val col = "col".voidTag[ReactOutput]
+  val col = "col".voidTag[ReactElement]
   /**
    * The table body.
    *
    *  MDN
    */
-  val tbody = "tbody".tag[ReactOutput]
+  val tbody = "tbody".tag[ReactElement]
   /**
    * The table headers.
    *
    *  MDN
    */
-  val thead = "thead".tag[ReactOutput]
+  val thead = "thead".tag[ReactElement]
   /**
    * The table footer.
    *
    *  MDN
    */
-  val tfoot = "tfoot".tag[ReactOutput]
+  val tfoot = "tfoot".tag[ReactElement]
   /**
    * A single row in a table.
    *
    *  MDN
    */
-  val tr = "tr".tag[ReactOutput]
+  val tr = "tr".tag[ReactElement]
   /**
    * A single cell in a table.
    *
    *  MDN
    */
-  val td = "td".tag[ReactOutput]
+  val td = "td".tag[ReactElement]
   /**
    * A header cell in a table.
    *
    *  MDN
    */
-  val th = "th".tag[ReactOutput]
+  val th = "th".tag[ReactElement]
 
   // Forms
   /**
@@ -463,65 +464,65 @@ trait ReactTags extends generic.Tags[VDomBuilder, ReactOutput, ReactFragT]{
    *
    *  MDN
    */
-  val form = "form".tag[ReactOutput]
+  val form = "form".tag[ReactElement]
   /**
    * A set of fields.
    *
    *  MDN
    */
-  val fieldset = "fieldset".tag[ReactOutput]
+  val fieldset = "fieldset".tag[ReactElement]
   /**
    * The caption for a fieldset.
    *
    *  MDN
    */
-  val legend = "legend".tag[ReactOutput]
+  val legend = "legend".tag[ReactElement]
   /**
    * The caption of a single field
    *
    *  MDN
    */
-  val label = "label".tag[ReactOutput]
+  val label = "label".tag[ReactElement]
   /**
    * A typed data field allowing the user to input data.
    *
    *  MDN
    */
-  val input = "input".voidTag[ReactOutput]
+  val input = "input".voidTag[ReactElement]
   /**
    * A button
    *
    *  MDN
    */
-  val button = "button".tag[ReactOutput]
+  val button = "button".tag[ReactElement]
   /**
    * A control that allows the user to select one of a set of options.
    *
    *  MDN
    */
-  val select = "select".tag[ReactOutput]
+  val select = "select".tag[ReactElement]
   /**
    * A set of predefined options for other controls.
    *
    *  MDN
    */
-  val datalist = "datalist".tag[ReactOutput]
+  val datalist = "datalist".tag[ReactElement]
   /**
    * A set of options, logically grouped.
    *
    *  MDN
    */
-  val optgroup = "optgroup".tag[ReactOutput]
+  val optgroup = "optgroup".tag[ReactElement]
   /**
    * An option in a select element.
    *
    *  MDN
    */
-  val option = "option".tag[ReactOutput]
+  val option = "option".tag[ReactElement]
   /**
    * A multiline text edit control.
    *
    *  MDN
    */
-  val textarea = "textarea".tag[ReactOutput]
+  val textarea = "textarea".tag[ReactElement]
 }