+++ core/src/main/scala/japgolly/scalajs/react/vdom/ReactVDom.scala
@@ -8,8 +8,8 @@ import scalatags.generic._
 import japgolly.scalajs.react._
 
 object ReactVDom
-    extends Bundle[VDomBuilder, ReactOutput, ReactFragT]
-    with Aliases[VDomBuilder, ReactOutput, ReactFragT] {
+    extends Bundle[VDomBuilder, ReactElement, ReactFragT]
+    with Aliases[VDomBuilder, ReactElement, ReactFragT] {
 
   object attrs extends ReactVDom.Cap with Attrs with ExtraAttrs
   object tags extends ReactVDom.Cap with ReactTags
@@ -40,9 +40,9 @@ object ReactVDom
     object * extends Cap with Attrs with Styles
   }
 
-  override type Tag = ReactVDom.TypedTag[ReactOutput]
+  override type Tag = ReactVDom.TypedTag[ReactElement]
 
-  sealed trait Aggregate extends generic.Aggregate[VDomBuilder, ReactOutput, ReactFragT] {
+  sealed trait Aggregate extends generic.Aggregate[VDomBuilder, ReactElement, ReactFragT] {
     override def genericAttr[T] = new GenericAttr[T](a => a.toString)
     override def genericStyle[T] = new GenericStyle[T]
 
@@ -70,28 +70,12 @@ object ReactVDom
     implicit val jsObjAttr = new GenericAttr[js.Object](f => f)
     implicit def reactRefAttr[T <: Ref[_]] = new GenericAttr[T](_.name)
 
-    implicit def modifierFromVDom(c: VDom): Modifier = new Modifier {
+    implicit def reactNodeAsDomChild[T <% ReactNode](c: T): Modifier = new Modifier {
       override def applyTo(t: VDomBuilder): Unit = t.appendChild(c)
     }
 
-    implicit def modifierFromPropsChildren(c: PropsChildren): Modifier = new Modifier {
-      override def applyTo(t: VDomBuilder): Unit = t.appendChild(c)
-    }
-
-    implicit def modifierFromSeqRCU_(cs: Seq[ReactComponentU_]): Modifier = new Modifier {
-      override def applyTo(t: VDomBuilder): Unit = t.appendChild(cs.asJsArray)
-    }
-
-    implicit def modifierFromArrVdom[T <% VDom](c: js.Array[T]): Modifier = new Modifier {
-      override def applyTo(t: VDomBuilder): Unit = t.appendChild(c)
-    }
-
-    implicit def vdomFromArrVdom[T <% VDom](cs: js.Array[T]): VDom = cs.asInstanceOf[VDom]
-    implicit def vdomFromSeqVdom[T <% VDom](cs: Seq[T])     : VDom = cs.asJsArray
-    implicit def vdomFromSeqTag            (cs: Seq[Tag])   : VDom = cs.toJsArray
-
-    @inline final implicit def autoRender(t: Tag) = t.render
-    @inline final implicit def autoRenderS(s: Seq[Tag]) = s.map(_.render)
+    @inline implicit def autoRender(t: Tag)      : ReactElement      = t.render
+    @inline implicit def autoRenderS(t: Seq[Tag]): Seq[ReactElement] = t.map(_.render)
 
     final def compositeAttr[A](k: Attr, f: (A, List[A]) => A, e: => Modifier = EmptyTag) =
       new CompositeAttr(k, f, e)
@@ -120,17 +104,17 @@ object ReactVDom
   }
 
   trait Cap extends Util { self =>
-    type ConcreteHtmlTag[T <: ReactOutput] = TypedTag[T]
+    type ConcreteHtmlTag[T <: ReactElement] = TypedTag[T]
 
     protected[this] implicit val stringAttrX: AttrValue[String] = new GenericAttr[String](s => s)
     protected[this] implicit val stringStyleX: StyleValue[String] = new GenericStyle[String]
 
-    def makeAbstractTypedTag[T <: ReactOutput](tag: String, void: Boolean, namespaceConfig: Namespace): TypedTag[T] =
+    def makeAbstractTypedTag[T <: ReactElement](tag: String, void: Boolean, namespaceConfig: Namespace): TypedTag[T] =
       TypedTag(tag, Nil, void, namespaceConfig)
 
     implicit class SeqFrag[A <% Frag](xs: Seq[A]) extends Frag{
       def applyTo(t: VDomBuilder): Unit = xs.foreach(_.applyTo(t))
-      def render: ReactOutput = {
+      def render: ReactElement = {
         val b = new VDomBuilder()
 
         applyTo(b)
@@ -164,7 +148,7 @@ object ReactVDom
     }
   }
 
-  final case class TypedTag[+Output <: ReactOutput](tag: String = "",
+  final case class TypedTag[+Output <: ReactElement](tag: String = "",
                                                     modifiers: List[Seq[Modifier]],
                                                     void: Boolean = false,
                                                     namespace: Namespace)
@@ -234,10 +218,5 @@ object ReactVDom
   implicit final class ReactBoolExt(val a: Boolean) extends AnyVal {
     @inline def &&(m: => Modifier): Modifier = if (a) m else EmptyTag
     // @inline def :=>[V](v: => V): Option[V] = if (a) Some(v) else None
-  }
-
-  implicit final class ArrayChildrenExt[A](val as: Seq[A]) extends AnyVal {
-    @inline def asJsArray = js.Array(as: _*)
-    @inline def toJsArray(implicit ev: A =:= TypedTag[ReactOutput]) = js.Array(as.map(_.render): _*)
   }
 }