+++ HISTORY.md
@@ -3,7 +3,28 @@ History
 
 ### 0.6.0 (unreleased)
 
-* Changed `ReactS.callback` method signatures from `(c)(a)` to `(a,c)`. **This breaks backwards compatibility.**
+This release brings scalajs-react in line with React 0.12.
+**React version 0.12.0 or later is now required.**
+
+Changes in React 0.12 can be read here:
+*  https://github.com/facebook/react/releases/tag/v0.12.0
+*  http://facebook.github.io/react/blog/2014/10/28/react-v0.12.html
+*  http://facebook.github.io/react/docs/glossary.html
+
+In addition to React API changes...
+* Deprecated `ReactOutput` and `VDom` in favour of `ReactElement` or in rare cases, `ReactNode`. (*[glossary](http://facebook.github.io/react/docs/glossary.html)*)
+* `.asJsArray: Seq[A] → JArray[A]` renamed to `toJsArray`
+* `.toJsArray: Seq[A] → JArray[ReactElement]` is no longer needed.
+* Renamed `ComponentSpec` to `ReactComponentSpec`. *(Internal. Extremely unlikely anyone using it directly.)*
+* Changed signatures of `ReactS.callback` and brethren from `(c)(a)` to `(a,c)`.
+
+Here are a few commands to ease migration.
+```
+find -name '*.scala' -exec perl -pi -e 's/(?<!\w)(vdom\.)?ReactOutput(?!\w)/ReactElement/g' {} +
+find -name '*.scala' -exec perl -pi -e 's/(?<!\w)VDom(?!\w)/ReactElement/g' {} +
+find -name '*.scala' -exec perl -pi -e 's/(?<!\w)asJsArray(?!\w)/toJsArray/g' {} + // careful...
+find -name '*.scala' -exec perl -pi -e 's/(?<=[ .]render)Component//g' {} +
+```
 
 ### 0.5.2 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.1...v0.5.2))